Renaming and separating classes
