+++ Anima/quantitative-mri/estimation/animaB1GMMRelaxometryCostFunction.cxx
@@ -1,4 +1,5 @@
 #include <animaB1GMMRelaxometryCostFunction.h>
+#include <animaB1GMMDistributionIntegrand.h>
 #include <animaBaseTensorTools.h>
 #include <animaGaussLaguerreQuadrature.h>
 
@@ -6,16 +7,6 @@
 
 namespace anima
 {
-
-double B1GMMDistributionIntegrand::operator() (double const t)
-{
-    if (m_EPGVectors.find(t) == m_EPGVectors.end())
-        m_EPGVectors.insert(std::make_pair(t,m_EPGSimulator.GetValue(m_T1Value, t, m_FlipAngle, 1.0)));
-
-    double gaussianExponent = (t - m_GaussianMean) * (t - m_GaussianMean) / (2.0 * m_GaussianVariance);
-
-    return m_EPGVectors[t][m_EchoNumber] * std::exp(- gaussianExponent) / (std::sqrt(2.0 * M_PI * m_GaussianVariance));
-}
 
 B1GMMRelaxometryCostFunction::MeasureType
 B1GMMRelaxometryCostFunction::GetValue(const ParametersType & parameters) const
